+++ src/baminsurances/data/HomeInsurance.java
@@ -13,8 +13,8 @@ public class HomeInsurance extends Insurance {
     private String buildingMaterial;
     private String standard;
     private int squareMetres;
-    private long homeAmount;
-    private long contentsAmount;
+    private int homeAmount;
+    private int contentsAmount;
     
     /**
      * Creates a new home insurance with the given values.
@@ -36,7 +36,7 @@ public class HomeInsurance extends Insurance {
     public HomeInsurance(Employee employee, int premium, int amount,
             String terms, String streetAddress, String zipCode,
             HomeType homeType, String buildingMaterial, String standard,
-            int squareMetres, long homeAmount, long contentsAmount) {
+            int squareMetres, int homeAmount, int contentsAmount) {
         super(employee, premium, amount, terms);
         setStreetAddress(streetAddress);
         setZipCode(zipCode);
@@ -171,7 +171,7 @@ public class HomeInsurance extends Insurance {
      * 
      * @return the home amount for this home insurance
      */
-    public long getHomeAmount() {
+    public int getHomeAmount() {
         return homeAmount;
     }
 
@@ -180,7 +180,7 @@ public class HomeInsurance extends Insurance {
      * 
      * @param homeAmount the new home amount
      */
-    public void setHomeAmount(long homeAmount) {
+    public void setHomeAmount(int homeAmount) {
         this.homeAmount = homeAmount;
     }
 
@@ -189,7 +189,7 @@ public class HomeInsurance extends Insurance {
      * 
      * @return the contents amount for this home insurance
      */
-    public long getContentsAmount() {
+    public int getContentsAmount() {
         return contentsAmount;
     }
 
@@ -198,7 +198,7 @@ public class HomeInsurance extends Insurance {
      * 
      * @param contentsAmount the new home amount
      */
-    public void setContentsAmount(long contentsAmount) {
+    public void setContentsAmount(int contentsAmount) {
         this.contentsAmount = contentsAmount;
     }
